notifications/requirements.txt
kafka-python==2.0.2
pika==1.2.0
